+++ src/Doctrine/Migration/FileBasedMigration.php
@@ -167,6 +167,7 @@ abstract class FileBasedMigration extends AbstractMigration
      */
     private function prepare(array $files)
     {
+        $this->queries = [];
         foreach ($files as $file) {
             assert('is_readable($file)');
             foreach (Parser::extractSql(file_get_contents($file)) as $sql) {