+++ tests/TestCase.php
@@ -10,6 +10,14 @@ namespace OctoLab\Common;
 abstract class TestCase extends \PHPUnit_Framework_TestCase
 {
     /**
+     * @return string
+     */
+    protected function getAppDir(): string
+    {
+        return sprintf('%s/fixtures/app', __DIR__);
+    }
+
+    /**
      * @param string $config
      * @param string $extension
      *
@@ -17,14 +25,24 @@ abstract class TestCase extends \PHPUnit_Framework_TestCase
      */
     protected function getConfigPath(string $config = 'config', string $extension = 'yml'): string
     {
-        return sprintf('%s/app/config/%s.%s', __DIR__, $config, $extension);
+        return sprintf('%s/fixtures/config/%s.%s', __DIR__, $config, $extension);
     }
 
     /**
      * @return string
      */
-    protected function getAppDir(): string
+    protected function getMigrationDir(): string
     {
-        return realpath(__DIR__ . '/app');
+        return sprintf('%s/fixtures/migrations', __DIR__);
+    }
+
+    /**
+     * @param string $migration
+     *
+     * @return string
+     */
+    protected function getMigrationPath(string $migration): string
+    {
+        return sprintf('%s/%s', $this->getMigrationDir(), $migration);
     }
 }