+++ tests/fixtures/migrations/2/ISSUE-29/upgrade.sql
@@ -1,0 +1,5 @@
+CREATE TABLE IF NOT EXISTS test (
+  id INT,
+  title VARCHAR(8) NOT NULL,
+  PRIMARY KEY (id)
+);